ml-service-implementations/ads-ml-service/requirements.txt
#  Dependences for web service

## Data structure
numpy
pandas
pydantic

## Web interface
pyyaml                    # server configurations
python-multipart          # file service
fastapi
uvicorn
gunicorn
expiringdict
sqlalchemy


# Load pre-trained model
requests~=2.23.0